[#8]: refactor: 리뷰 서비스 통합
